refactor-commit -SumMultiplesOf3And5Below, Sum, SumDivisibleBy

+++ ProjectEuler/Problem1.h
@@ -4,7 +4,7 @@
 
 class PROJECTEULER_API Problem1 {
 public:
-    static unsigned long SumMultiplesOf3And5Below(int);
-    static long Sum(long);
-    static long SumDivisibleBy(int, long);
+    static unsigned long SumMultiplesOf3And5Below(unsigned long);
+    static unsigned long Sum(unsigned long);
+    static unsigned long SumDivisibleBy(unsigned long, unsigned long);
 };
